# Conflicts:
#	configure
#	libavcodec/Makefile
#	libavcodec/mediacodec.c
#	libavcodec/mediacodec.h
#	libavcodec/mediacodec_surface.c
#	libavcodec/mediacodec_wrapper.c
#	libavcodec/mediacodecdec.c
#	libavformat/Makefile
#	libavformat/cache.c
#	libavformat/libavformat.v
#	libavutil/pixfmt.h
#	libswscale/arm/Makefile
#	libswscale/arm/swscale_unscaled.c
#	libswscale/swscale_unscaled.c

+++ libavcodec/vc2enc_dwt.h
@@ -25,9 +25,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef int16_t dwtcoef;
-
-/* Only Deslauriers-Dubuc (9,7) and LeGall (5,3) supported! */
+typedef int32_t dwtcoef;
 
 enum VC2TransformType {
     VC2_TRANSFORM_9_7    = 0,   /* Deslauriers-Dubuc (9,7)  */
